expenses/requirements.txt
Flask
mysql-connector-python
matplotlib
gunicorn